Move Match rules into own module.
